requirements.txt
accelerate==0.29.3
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
asttokens==2.4.1
asyncio==3.4.3
attrs==23.2.0
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.5
decorator==5.1.1
diffusers==0.11.1
disutils==1.4.32.post2
executing==2.0.1
filelock==3.14.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.23.0
idna==3.7
importlib_metadata==7.1.0
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonpatch==1.33
jsonpickle==3.2.1
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain-community==0.0.36
langchain-core==0.1.50
langsmith==0.1.54
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
orjson==3.10.3
packaging==23.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyvis==0.3.2
PyYAML==6.0.1
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.30
st-click-detector==0.1.3
stack-data==0.6.3
streamlit==1.34.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.40.1
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
zipp==3.18.1
